+++ aerospike/datadog_checks/aerospike/config_models/shared.py
@@ -19,12 +19,24 @@ from datadog_checks.base.utils.models import validation
 from . import defaults, validators
 
 
+class Proxy(BaseModel):
+    class Config:
+        allow_mutation = False
+
+    http: Optional[str]
+    https: Optional[str]
+    no_proxy: Optional[Sequence[str]]
+
+
 class SharedConfig(BaseModel):
     class Config:
         allow_mutation = False
 
     mappings: Optional[Sequence[str]]
+    proxy: Optional[Proxy]
     service: Optional[str]
+    skip_proxy: Optional[bool]
+    timeout: Optional[float]
 
     @root_validator(pre=True)
     def _initial_validation(cls, values):