+++ avi_vantage/datadog_checks/avi_vantage/config_models/instance.py
@@ -127,6 +127,7 @@ class InstanceConfig(BaseModel):
     service: Optional[str]
     share_labels: Optional[Mapping[str, Union[bool, ShareLabel]]]
     skip_proxy: Optional[bool]
+    tag_by_endpoint: Optional[bool]
     tags: Optional[Sequence[str]]
     telemetry: Optional[bool]
     timeout: Optional[float]