+++ haproxy/datadog_checks/haproxy/config_models/defaults.py
@@ -62,6 +62,10 @@ def instance_bearer_token_path(field, value):
     return get_default_field_value(field, value)
 
 
+def instance_bearer_token_refresh_interval(field, value):
+    return 60
+
+
 def instance_cache_metric_wildcards(field, value):
     return True
 
@@ -340,6 +344,10 @@ def instance_startup_grace_seconds(field, value):
 
 def instance_status_check(field, value):
     return False
+
+
+def instance_tag_by_endpoint(field, value):
+    return True
 
 
 def instance_tag_service_check_by_host(field, value):