+++ vault/CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG - Vault
 
+## 3.2.0 / 2022-05-11
+
+* [Added] Add `vault.replication.wal.gc.*` metrics. See [#11984](https://github.com/DataDog/integrations-core/pull/11984).
+* [Fixed] Fix metric naming. See [#11847](https://github.com/DataDog/integrations-core/pull/11847).
+
 ## 3.1.0 / 2022-04-05
 
 * [Added] Add metric_patterns options to filter all metric submission by a list of regexes. See [#11695](https://github.com/DataDog/integrations-core/pull/11695).